+++ CHANGELOG.md
@@ -1,7 +1,12 @@
-# Changelog
+# Power Scale Changelog
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.20.1-0.1.0] - Unreleased
 - First release
+### Added
+- Difficulty system (incomplete, but functioning)
+- Scaling Attributes system
+- Alchemy Set with brews and tonics
+- Commands